+++ src/main/java/subway/controller/SubwayController.java
@@ -13,6 +13,7 @@ import java.util.List;
 
 public class SubwayController {
     private WeightedMultigraph<String, DefaultWeightedEdge> distanceGraph;
+    private WeightedMultigraph<String, DefaultWeightedEdge> timeGraph;
 
     public void initPrimary() {
         Arrays.asList("교대역", "강남역", "역삼역", "남부터미널역", "양재역", "양재시민의숲역", "매봉역")
@@ -26,14 +27,29 @@ public class SubwayController {
         EdgeRepository.addEdge(new Edge("양재역", "양재시민의숲역", 10, 3));
     }
 
-    public void initDistanceGraph() {
+    public void initGraph() {
+        initDistanceGraph();
+        initTimeGraph();
+    }
+    
+    private void initDistanceGraph() {
         this.distanceGraph = new WeightedMultigraph(DefaultWeightedEdge.class);
         StationRepository.stations().forEach(station -> distanceGraph.addVertex(station.getName()));
         EdgeRepository.edges().forEach(edge -> distanceGraph.setEdgeWeight(distanceGraph.addEdge(edge.getFrom(), edge.getTo()), edge.getDistance()));
     }
+    
+    private void initTimeGraph() {
+        this.timeGraph = new WeightedMultigraph(DefaultWeightedEdge.class);
+        StationRepository.stations().forEach(station -> timeGraph.addVertex(station.getName()));
+        EdgeRepository.edges().forEach(edge -> timeGraph.setEdgeWeight(timeGraph.addEdge(edge.getFrom(), edge.getTo()), edge.getTime()));
+    }
 
-    private List<String> getShortestPathDistance(String from, String to) {
+    public List<String> getShortestPathDistance(String from, String to) {
         return getShortestPath(from, to, this.distanceGraph);
+    }
+    
+    public List<String> getShortestPathTime(String from, String to) {
+        return getShortestPath(from, to, this.timeGraph);
     }
 
     private List<String> getShortestPath(String from, String to, WeightedMultigraph<String, DefaultWeightedEdge> graph) {